New Version
Switch and Meter were moved to seperate files with structs.
