+++ meter/shelly/switch.go
@@ -1,0 +1,148 @@
+package shelly
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+)
+
+type ShellySwitch struct {
+	Conn *Connection
+}
+
+func NewShellySwitch(conn *Connection) *ShellySwitch {
+	shellySwitch := &ShellySwitch{
+		Conn: conn,
+	}
+
+	return shellySwitch
+}
+
+// CurrentPower implements the api.Meter interface
+func (shellySwitch *ShellySwitch) CurrentPower() (float64, error) {
+	var power float64
+	d := shellySwitch.Conn
+
+	switch d.gen {
+	case 0, 1:
+		var res Gen1StatusResponse
+		uri := fmt.Sprintf("%s/status", d.uri)
+		if err := d.GetJSON(uri, &res); err != nil {
+			return 0, err
+		}
+
+		switch {
+		case d.channel < len(res.Meters):
+			power = res.Meters[d.channel].Power
+		case d.channel < len(res.EMeters):
+			power = res.EMeters[d.channel].Power
+		default:
+			return 0, errors.New("invalid channel, missing power meter")
+		}
+
+	default:
+		var res Gen2StatusResponse
+		if err := d.execGen2Cmd("Shelly.GetStatus", false, &res); err != nil {
+			return 0, err
+		}
+
+		switch d.channel {
+		case 1:
+			power = res.Switch1.Apower
+		case 2:
+			power = res.Switch2.Apower
+		default:
+			power = res.Switch0.Apower
+		}
+	}
+
+	return power, nil
+}
+
+// Enabled implements the api.Charger interface
+func (shellySwitch *ShellySwitch) Enabled() (bool, error) {
+	d := shellySwitch.Conn
+	switch d.gen {
+	case 0, 1:
+		var res Gen1SwitchResponse
+		uri := fmt.Sprintf("%s/relay/%d", d.uri, d.channel)
+		err := d.GetJSON(uri, &res)
+		return res.Ison, err
+
+	default:
+		var res Gen2SwitchResponse
+		err := d.execGen2Cmd("Switch.GetStatus", false, &res)
+		return res.Output, err
+	}
+}
+
+// Enable implements the api.Charger interface
+func (shellySwitch *ShellySwitch) Enable(enable bool) error {
+	var err error
+	onoff := map[bool]string{true: "on", false: "off"}
+	d := shellySwitch.Conn
+
+	switch d.gen {
+	case 0, 1:
+		var res Gen1SwitchResponse
+		uri := fmt.Sprintf("%s/relay/%d?turn=%s", d.uri, d.channel, onoff[enable])
+		err = d.GetJSON(uri, &res)
+
+	default:
+		var res Gen2SwitchResponse
+		err = d.execGen2Cmd("Switch.Set", enable, &res)
+	}
+
+	return err
+}
+
+// TotalEnergy implements the api.Meter interface
+func (shellySwitch *ShellySwitch) TotalEnergy() (float64, error) {
+	var energy float64
+	d := shellySwitch.Conn
+	switch d.gen {
+	case 0, 1:
+		var res Gen1StatusResponse
+		uri := fmt.Sprintf("%s/status", d.uri)
+		if err := d.GetJSON(uri, &res); err != nil {
+			return 0, err
+		}
+
+		switch {
+		case d.channel < len(res.Meters):
+			energy = res.Meters[d.channel].Total
+		case d.channel < len(res.EMeters):
+			energy = res.EMeters[d.channel].Total
+		default:
+			return 0, errors.New("invalid channel, missing power meter")
+		}
+
+		energy = GetGen1Energy(d.devicetype, energy)
+
+	default:
+		var res Gen2StatusResponse
+		if err := d.execGen2Cmd("Shelly.GetStatus", false, &res); err != nil {
+			return 0, err
+		}
+
+		switch d.channel {
+		case 1:
+			energy = res.Switch1.Aenergy.Total
+		case 2:
+			energy = res.Switch2.Aenergy.Total
+		default:
+			energy = res.Switch0.Aenergy.Total
+		}
+	}
+
+	return energy / 1000, nil
+}
+
+// GetGen1Energy in kWh
+func GetGen1Energy(devicetype string, energy float64) float64 {
+	// Gen 1 Shelly EM devices are providing Watt hours, Shelly EM devices are providing Watt minutes
+	if !strings.Contains(devicetype, "EM") {
+		energy = energy / 60
+	}
+	return energy
+}
