+++ templates/definition/meter/shelly-1pm.yaml
@@ -11,13 +11,10 @@ params:
     default: 192.0.2.2
   - name: user
   - name: password
-    mask: true
   - name: channel
-    default: 0
 render: |
   type: shelly
   uri: http://{{ .host }}  # shelly device ip address (local)
   {{ if ne .user "" }}user: {{ .user }}{{ end }}
   {{ if ne .password "" }}password: {{ .password }}{{ end }}
   channel: {{ .channel }}  # shelly device relay channel
-  deviceType: switch #shelly device type
