+++ templates/definition/meter/shelly-pro-3em.yaml
@@ -10,13 +10,9 @@ params:
     default: 192.0.2.2
   - name: user
   - name: password
-    mask: true
-  - name: channel
-    default: 0
 render: |
-  type: shelly
+  type: shelly-energy-meter
   uri: http://{{ .host }}  # shelly device ip address (local)
   {{ if ne .user "" }}user: {{ .user }}{{ end }}
   {{ if ne .password "" }}password: {{ .password }}{{ end }}
-  channel: {{ .channel }}  # shelly device relay channel
-  deviceType: energyMeter #shelly device type
+  channel: 0  # shelly device relay channel
